address review (nik): move expiry email models to tenants/models.py
